+++ src/components/Button.h
@@ -20,7 +20,7 @@ public:
         _platform->pinMode(_pin, INPUT_PULLUP);
     }
 
-    void update(unsigned int delta) override {
+    void update(unsigned long delta) override {
         bool pressed = _platform->digitalRead(_pin) == LOW;
 
         if (!_pressed && pressed) {
