+++ src/interval/Interval.h
@@ -9,11 +9,11 @@ public:
      * Update the duration of the _interval.
      * @param duration
      */
-    virtual void setDuration(unsigned int duration) = 0;
+    virtual void setDuration(unsigned long duration) = 0;
 
     virtual void reset() = 0;
 
-    virtual unsigned int getDuration() = 0;
+    virtual unsigned long getDuration() = 0;
 };
 
 #endif //CUBE_ARDUINO_SDK_INTERVAL_H
